+++ sampledb/frontend/users/authentication.py
@@ -3,12 +3,14 @@
 
 """
 
+import datetime
+
 import flask
 import flask_login
 from flask_babel import _, lazy_gettext, refresh
 
 from .. import frontend
-from ...logic.authentication import login
+from ...logic.authentication import login, get_active_two_factor_authentication_method
 from ...logic.users import get_user
 from ...frontend.users_forms import SigninForm, SignoutForm
 from ... import login_manager
@@ -70,19 +72,15 @@ def _sign_in_impl(is_for_refresh):
         username = username.lower()
         password = form.password.data
         user = login(username, password)
-        if user and not user.is_active:
-            flask_login.logout_user()
-            flask.flash(_('Your user account has been deactivated. Please contact an administrator.'), 'error')
-            return flask.redirect(flask.url_for('frontend.sign_in'))
-        if is_for_refresh:
-            if user and user == flask_login.current_user:
-                flask_login.confirm_login()
-                return _redirect_to_next_url()
-        else:
-            if user:
-                remember_me = form.remember_me.data
-                flask_login.login_user(user, remember=remember_me)
-                refresh()
+        if user:
+            if not user.is_active:
+                flask_login.logout_user()
+                flask.flash(_('Your user account has been deactivated. Please contact an administrator.'), 'error')
+                return flask.redirect(flask.url_for('frontend.sign_in'))
+            if is_for_refresh and user != flask_login.current_user:
+                user = None
+        if user:
+            if not is_for_refresh:
                 if user.is_admin and password == 'password':
                     flask.flash(
                         lazy_gettext(
@@ -93,11 +91,37 @@ def _sign_in_impl(is_for_refresh):
                         ),
                         'warning'
                     )
-                return _redirect_to_next_url()
+            two_factor_authentication_method = get_active_two_factor_authentication_method(user.id)
+            if two_factor_authentication_method is not None:
+                if two_factor_authentication_method.data.get('type') == 'totp':
+                    flask.session['confirm_data'] = {
+                        'reason': 'login',
+                        'user_id': two_factor_authentication_method.user_id,
+                        'method_id': two_factor_authentication_method.id,
+                        'is_for_refresh': is_for_refresh,
+                        'remember_me': form.remember_me.data,
+                        'expiration_datetime': (datetime.datetime.utcnow() + datetime.timedelta(minutes=5)).strftime('%Y-%m-%d %H:%M:%S')
+                    }
+                    return flask.redirect(flask.url_for('.confirm_totp_two_factor_authentication'))
+                return flask.render_template('two_factor_authentication/unsupported_method.html')
+            return complete_sign_in(user, is_for_refresh, form.remember_me.data)
         has_errors = True
     elif form.errors:
         has_errors = True
     return flask.render_template('sign_in.html', form=form, is_for_refresh=is_for_refresh, has_errors=has_errors)
+
+
+def complete_sign_in(user, is_for_refresh, remember_me):
+    if not user:
+        flask.flash(_('Please sign in again.'), 'error')
+        flask_login.logout_user()
+        return flask.redirect(flask.url_for('.index'))
+    if is_for_refresh:
+        flask_login.confirm_login()
+    else:
+        flask_login.login_user(user, remember=remember_me)
+        refresh()
+    return _redirect_to_next_url()
 
 
 @frontend.route('/users/me/sign_in', methods=['GET', 'POST'])